Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI Writing Assistant Backend"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "sqlmodel[asyncio]>=0.0.24",
    "psycopg2-binary>=2.9.10",
    "asyncpg>=0.30.0",
    "pyjwt>=2.10.1",
    "passlib==1.7.4",
    "bcrypt==3.2.0",
    "python-dotenv>=0.21.0",
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "slowapi>=0.1.9",
    "google-auth>=2.40.3",
]
